feat: refactor postbuild script
