+++ packages/create-abell/package.json
@@ -10,7 +10,8 @@
   "scripts": {
     "test": "vitest",
     "test:once": "vitest run",
-    "build": "tsc && node scripts/post-build.js",
+    "build": "tsc",
+    "postbuild": "node ../../scripts/post-build.js create-abell",
     "dev": "nodemon --exec \"yarn build\" --watch src --watch templates -e js,ts,abell,css",
     "clean-scaffolds": "node scripts/clean-scaffolds.js",
     "scaffold": "npm run clean-scaffolds && cd scaffold-dir && node ../dist/bin.js",
